refactor: vlan details MAASENG-1828 (#5005)
