+++ src/app/subnets/views/VLANDetails/VLANDeleteForm/VLANDeleteForm.test.tsx
