+++ src/app/subnets/views/VLANDetails/VLANDeleteForm/VLANDeleteForm.tsx
