+++ src/app/subnets/views/VLANDetails/VLANDeleteForm/index.ts
